Added a test.
The new approach is less testable than the old approach, which is
frustrating. On the other hand, there is much less to actually test,
because "make -npBr" is doing the heavy lifting of actually processing
the Makefile, and graphviz-python is making the graph. So I think the
trade off was worth it.